fix(ui): move custom svg files
