+++ packages/ui-components/src/deprecated_js/Icon/Icon.component.js
@@ -22,7 +22,7 @@ import ChevronLeft from "@material-design-icons/svg/outlined/chevron_left.svg"
 import ChevronRight from "@material-design-icons/svg/outlined/chevron_right.svg"
 import Close from "@material-design-icons/svg/filled/close.svg"
 import ContentCopy from "@material-design-icons/svg/outlined/content_copy.svg"
-import Danger from "./icons/juno-danger.svg"
+import Danger from "../../components/Icon/icons/juno-danger.svg"
 import Dangerous from "@material-design-icons/svg/filled/dangerous.svg"
 import Download from "@material-design-icons/svg/filled/download.svg"
 import DeleteForever from "@material-design-icons/svg/filled/delete_forever.svg"
@@ -37,7 +37,7 @@ import ExpandMore from "@material-design-icons/svg/outlined/expand_more.svg"
 import FilterAlt from "@material-design-icons/svg/filled/filter_alt.svg"
 import Forum from "@material-design-icons/svg/filled/forum.svg"
 import Help from "@material-design-icons/svg/filled/help.svg"
-import Home from "./icons/home_sharp.svg"
+import Home from "../../components/Icon/icons/home_sharp.svg"
 import Info from "@material-design-icons/svg/filled/info.svg"
 import Comment from "@material-design-icons/svg/filled/comment.svg"
 import ManageAccounts from "@material-design-icons/svg/filled/manage_accounts.svg"
@@ -47,13 +47,13 @@ import NightsStay from "@material-design-icons/svg/outlined/nights_stay.svg"
 import NotificationsOff from "@material-design-icons/svg/outlined/notifications_off.svg"
 import OpenInBrowser from "@material-design-icons/svg/outlined/open_in_browser.svg"
 import OpenInNew from "@material-design-icons/svg/outlined/open_in_new.svg"
-import Place from "./icons/place.svg"
+import Place from "../../components/Icon/icons/place.svg"
 import Success from "@material-design-icons/svg/filled/check_box.svg"
 import Search from "@material-design-icons/svg/outlined/search.svg"
-import SeverityLow from "./icons/juno_severity_low.svg"
-import SeverityMedium from "./icons/juno_severity_medium.svg"
-import SeverityHigh from "./icons/juno_severity_high.svg"
-import SeverityCritical from "./icons/juno_severity_critical.svg"
+import SeverityLow from "../../components/Icon/icons/juno_severity_low.svg"
+import SeverityMedium from "../../components/Icon/icons/juno_severity_medium.svg"
+import SeverityHigh from "../../components/Icon/icons/juno_severity_high.svg"
+import SeverityCritical from "../../components/Icon/icons/juno_severity_critical.svg"
 import Upload from "@material-design-icons/svg/filled/upload.svg"
 import Warning from "@material-design-icons/svg/filled/warning.svg"
 import WBSunny from "@material-design-icons/svg/outlined/wb_sunny.svg"
